+++ contracts/mocks/GovernorCompMock.sol
@@ -2,34 +2,29 @@
 
 pragma solidity ^0.8.0;
 
-import "../governance/Governor.sol";
+import "../governance/extensions/GovernorSettings.sol";
 import "../governance/extensions/GovernorCountingSimple.sol";
 import "../governance/extensions/GovernorVotesComp.sol";
 
-contract GovernorCompMock is Governor, GovernorVotesComp, GovernorCountingSimple {
-    uint256 immutable _votingDelay;
-    uint256 immutable _votingPeriod;
-
+contract GovernorCompMock is GovernorSettings, GovernorVotesComp, GovernorCountingSimple {
     constructor(
         string memory name_,
         ERC20VotesComp token_,
         uint256 votingDelay_,
         uint256 votingPeriod_
-    ) Governor(name_) GovernorVotesComp(token_) {
-        _votingDelay = votingDelay_;
-        _votingPeriod = votingPeriod_;
-    }
-
-    function votingDelay() public view override returns (uint256) {
-        return _votingDelay;
-    }
-
-    function votingPeriod() public view override returns (uint256) {
-        return _votingPeriod;
-    }
+    ) Governor(name_) GovernorSettings(votingDelay_, votingPeriod_, 0) GovernorVotesComp(token_) {}
 
     function quorum(uint256) public pure override returns (uint256) {
         return 0;
+    }
+
+    function propose(
+        address[] memory targets,
+        uint256[] memory values,
+        bytes[] memory calldatas,
+        string memory description
+    ) public virtual override(Governor, GovernorProposalThreshold) returns (uint256) {
+        return super.propose(targets, values, calldatas, description);
     }
 
     function cancel(